Added Function to find neighbouring cells
call getNeighbours(int i) in the TerrainHandler, where i is the index of a cell. The function will return an array of integers referring to each neighbouring cell. a value of -1 means that the cell does not exist. This likely means that the cell is on the edge of the map. Function should be refactored later on for efficiency